[scramjet/core] proper document.referer emulation via worker side history tracking and client id url parameters
