+++ packages/scramjet/packages/core/src/fetch/body.ts
@@ -1,0 +1,81 @@
+import { BareResponse } from "@mercuryworkshop/proxy-transports";
+import {
+	BodyType,
+	ScramjetFetchHandler,
+	ScramjetFetchParsed,
+	ScramjetFetchRequest,
+} from ".";
+import {
+	isHtmlMimeType,
+	isJavascriptMimeType,
+	rewriteCss,
+	rewriteHtml,
+	rewriteJs,
+	rewriteWorkers,
+} from "@/shared";
+import { sniffEncoding } from "@/shared/sniffEncoding";
+
+export async function rewriteBody(
+	handler: ScramjetFetchHandler,
+	request: ScramjetFetchRequest,
+	parsed: ScramjetFetchParsed,
+	response: BareResponse
+): Promise<BodyType> {
+	switch (request.destination) {
+		case "iframe":
+		case "document":
+			if (isHtmlMimeType(response.headers.get("content-type") ?? "")) {
+				const buf = await response.arrayBuffer();
+				const bytes = new Uint8Array(buf);
+				const encoding = sniffEncoding(
+					bytes,
+					response.headers.get("content-type")
+				);
+				const htmlContent = new TextDecoder(encoding).decode(bytes);
+
+				console.log(parsed.trackedClient);
+				return rewriteHtml(htmlContent, handler.context, parsed.meta, {
+					loadScripts: true,
+					inline: true,
+					source: parsed.url.href,
+					headers: response.rawHeaders,
+					history: parsed.trackedClient!.history,
+				});
+			} else {
+				return response.body;
+			}
+		case "script": {
+			// do not attempt to rewrite a 404 response
+			if (response.ok) {
+				const ct = response.headers.get("content-type");
+				// don't rewrite invalid module scripts when the server declares a non-JS type
+				if (parsed.scriptType === "module" && ct && !isJavascriptMimeType(ct)) {
+					return response.body;
+				}
+
+				return rewriteJs(
+					new Uint8Array(await response.arrayBuffer()),
+					response.url,
+					handler.context,
+					parsed.meta,
+					parsed.scriptType === "module"
+				) as unknown as ArrayBuffer;
+			}
+			return response.body;
+		}
+		case "style":
+			return rewriteCss(await response.text(), handler.context, parsed.meta);
+		case "sharedworker":
+		case "worker":
+			return rewriteWorkers(
+				handler.context,
+				new Uint8Array(await response.arrayBuffer()),
+				// TODO: this takes a scriptType and rewritejs takes a bool..
+				parsed.scriptType,
+				response.url,
+				parsed.meta
+			);
+		default:
+			return response.body;
+	}
+}
