+++ packages/scramjet/packages/core/src/fetch/headers.ts
@@ -1,0 +1,198 @@
+import {
+	rewriteUrl,
+	ScramjetContext,
+	ScramjetHeaders,
+	unrewriteUrl,
+	URLMeta,
+} from "@/shared";
+import {
+	ScramjetFetchHandler,
+	ScramjetFetchParsed,
+	ScramjetFetchRequest,
+} from ".";
+import { RawHeaders } from "@mercuryworkshop/proxy-transports";
+
+/**
+ * Headers for security policy features that haven't been emulated yet
+ */
+const SEC_HEADERS = new Set([
+	"cross-origin-embedder-policy",
+	"cross-origin-opener-policy",
+	"cross-origin-resource-policy",
+	"content-security-policy",
+	"content-security-policy-report-only",
+	"expect-ct",
+	"feature-policy",
+	"origin-isolation",
+	"strict-transport-security",
+	"upgrade-insecure-requests",
+	"x-content-type-options",
+	"x-download-options",
+	"x-frame-options",
+	"x-permitted-cross-domain-policies",
+	"x-powered-by",
+	"x-xss-protection",
+	// This needs to be emulated, but for right now it isn't that important of a feature to be worried about
+	// https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Clear-Site-Data
+	"clear-site-data",
+]);
+
+/**
+ * Headers that are actually URLs that need to be rewritten
+ */
+const URL_HEADERS = new Set(["location", "content-location", "referer"]);
+
+function rewriteLinkHeader(
+	link: string,
+	context: ScramjetContext,
+	meta: URLMeta
+) {
+	return link.replace(/<([^>]+)>/gi, (_match, p1) => {
+		return `<${rewriteUrl(p1, context, meta)}>`;
+	});
+}
+
+export async function rewriteResponseHeaders(
+	handler: ScramjetFetchHandler,
+	request: ScramjetFetchRequest,
+	parsed: ScramjetFetchParsed,
+	rawHeaders: RawHeaders
+): Promise<ScramjetHeaders> {
+	const headers = ScramjetHeaders.fromRawHeaders(rawHeaders);
+
+	for (const cspHeader of SEC_HEADERS) {
+		headers.delete(cspHeader);
+	}
+
+	for (const urlHeader of URL_HEADERS) {
+		if (headers.has(urlHeader)) {
+			let url = headers.get(urlHeader)!;
+			let rewrittenUrl = rewriteUrl(url, handler.context, parsed.meta);
+			headers.set(urlHeader, rewrittenUrl);
+		}
+	}
+
+	if (headers.has("link")) {
+		let link = headers.get("link")!;
+		let rewritten = rewriteLinkHeader(link, handler.context, parsed.meta);
+		headers.set("link", rewritten);
+	}
+
+	if (headers.get("accept") === "text/event-stream") {
+		headers.set("content-type", "text/event-stream");
+	}
+
+	// scramjet runtime can use features that permissions-policy blocks
+	headers.delete("permissions-policy");
+
+	if (
+		handler.crossOriginIsolated &&
+		[
+			"document",
+			"iframe",
+			"worker",
+			"sharedworker",
+			"style",
+			"script",
+		].includes(request.destination)
+	) {
+		headers.set("Cross-Origin-Embedder-Policy", "require-corp");
+		headers.set("Cross-Origin-Opener-Policy", "same-origin");
+	}
+
+	if (request.destination === "document" || request.destination === "iframe") {
+		headers.set("Referrer-Policy", "unsafe-url");
+	}
+
+	return headers;
+}
+
+export function rewriteRequestHeaders(
+	request: ScramjetFetchRequest,
+	handler: ScramjetFetchHandler,
+	parsed: ScramjetFetchParsed
+): ScramjetHeaders {
+	const headers = request.initialHeaders.clone();
+
+	if (request.rawReferrer) {
+		const clientUrl = request.rawClientUrl || new URL(request.rawReferrer);
+		if (clientUrl.pathname.startsWith(handler.context.prefix.pathname)) {
+			let unrewritten = new URL(unrewriteUrl(clientUrl, handler.context));
+
+			const referer = createReferrerString(
+				unrewritten,
+				parsed.url,
+				request.rawReferrerPolicy
+			);
+			if (referer) headers.set("Referer", referer);
+		}
+	}
+
+	const cookies = handler.context.cookieJar.getCookies(parsed.url, false);
+
+	if (cookies.length) {
+		headers.set("Cookie", cookies);
+	}
+
+	return headers;
+}
+
+export function createReferrerString(
+	clientUrl: URL,
+	resource: URL,
+	policy: string | null
+): string {
+	policy ||= "strict-origin-when-cross-origin";
+	const originIsHttps = clientUrl.protocol === "https:";
+	const destIsHttps = resource.protocol === "https:";
+
+	// A "more private" request: https -> http
+	const isPotentialDowngrade = originIsHttps && !destIsHttps;
+
+	// Step 3: Determine if same-origin
+	const isSameOrigin =
+		clientUrl.protocol === resource.protocol &&
+		clientUrl.host === resource.host;
+
+	// Step 4: Strip referrer to just origin (scheme + host + port)
+	const referrerOrigin = clientUrl.origin; // e.g. "https://example.com"
+
+	const referrerUrl = new URL(clientUrl.href);
+	referrerUrl.hash = "";
+	const referrerUrlString = referrerUrl.href;
+
+	switch (policy) {
+		case "no-referrer":
+			return "";
+
+		case "no-referrer-when-downgrade":
+			if (isPotentialDowngrade) return "";
+			return referrerUrlString;
+
+		case "same-origin":
+			if (isSameOrigin) return referrerUrlString;
+			return "";
+
+		case "origin":
+			return referrerOrigin === "null" ? "" : referrerOrigin + "/";
+
+		case "strict-origin":
+			if (isPotentialDowngrade) return "";
+			return referrerOrigin === "null" ? "" : referrerOrigin + "/";
+
+		case "origin-when-cross-origin":
+			if (isSameOrigin) return referrerUrlString;
+			return referrerOrigin === "null" ? "" : referrerOrigin + "/";
+
+		case "strict-origin-when-cross-origin":
+			if (isSameOrigin) return referrerUrlString;
+			if (isPotentialDowngrade) return "";
+			return referrerOrigin === "null" ? "" : referrerOrigin + "/";
+
+		case "unsafe-url":
+			return referrerUrlString;
+
+		default:
+			return "";
+	}
+}
